Adding transaction method

+++ src/main/java/gov/sp/fatec/bookblog/service/CategoryService.java
@@ -1,17 +1,22 @@
 package gov.sp.fatec.bookblog.service;
 
+import gov.sp.fatec.bookblog.model.Book;
 import gov.sp.fatec.bookblog.model.Category;
 import gov.sp.fatec.bookblog.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.Optional;
 
 @Service
 public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    private BookService bookService;
 
     @PreAuthorize("permitAll()")
     public Iterable<Category> findAll() {
@@ -39,7 +44,17 @@ public class CategoryService {
     }
 
     @PreAuthorize("hasRole('ADMIN')")
+    @Transactional
     public void deleteById(final Long id) {
-        categoryRepository.deleteById(id);
+        Optional<Category> category = categoryRepository.findById(id);
+
+        if (category.isPresent()) {
+            for (Book book : category.get().getBooks()) {
+                book.getCategories().remove(category.get());
+
+                bookService.update(book);
+            }
+            categoryRepository.deleteById(id);
+        }
     }
 }
